usm: tests: classification: Move postgres classification test to USM linux file
